+++ packages/modeling/src/maths/utils/area.js
@@ -1,5 +1,8 @@
-/*
- * Calculate the area under the given points
+/**
+ * Calculate the area under the given points.
+ * @param {Array} points - list of 2D points
+ * @return {Number} area under the given points
+ * @alias module:modeling/maths/utils.area
  */
 const area = (points) => {
   let area = 0