+++ packages/modeling/src/maths/utils/interpolateBetween2DPointsForY.js
@@ -1,8 +1,11 @@
 /**
  * Get the X coordinate of a point with a certain Y coordinate, interpolated between two points.
  * Interpolation is robust even if the points have the same Y coordinate
+ * @param {vec2} point1
+ * @param {vec2} point2
+ * @param {Number} y
  * @return {Array} X and Y of interpolated point
- * @alias module:modeling/utils.interpolateBetween2DPointsForY
+ * @alias module:modeling/maths/utils.interpolateBetween2DPointsForY
  */
 const interpolateBetween2DPointsForY = (point1, point2, y) => {
   let f1 = y - point1[1]