+++ packages/modeling/src/maths/vec3/index.d.ts
@@ -30,7 +30,6 @@ export { default as squaredLength } from './squaredLength'
 export { default as subtract } from './subtract'
 export { default as toString } from './toString'
 export { default as transform } from './transform'
-export { default as unit } from './unit'
 
 export { default as Vec3 } from './type'
 export as namespace vec3