+++ packages/io/svg-serializer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/svg-serializer@2.2.4...@jscad/svg-serializer@2.2.5) (2021-09-27)
+
+**Note:** Version bump only for package @jscad/svg-serializer
+
+
+
+
+
 ## [2.2.4](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/svg-serializer@2.2.3...@jscad/svg-serializer@2.2.4) (2021-09-09)
 
 **Note:** Version bump only for package @jscad/svg-serializer